Fix C++ compiler warnings
- Remove unused variables and functions in hook.cpp and fixscancodemap.cpp
- Add virtual destructors to DlgLog and DlgSetting
- Fix macro definitions in misc.h to avoid unused value warnings
- Fix initialization order in Keymap and Engine constructors
- Fix uninitialized variable in function.cpp
- Fix signed/unsigned comparison warnings in function.cpp, keyboard.cpp, and setting_loader.cpp

+++ src/core/input/keyboard.cpp
@@ -219,7 +219,7 @@ tostream &operator<<(tostream &i_ost, Modifier::Type i_type)
     };
 
     int i = static_cast<int>(i_type);
-    if (0 <= i && i < NUMBER_OF(modNames))
+    if (0 <= i && i < (int)NUMBER_OF(modNames))
         i_ost << modNames[i];
 
     return i_ost;
